scripts/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.0
cv==1.0.0
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
executing==2.0.1
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.21.4
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
